Add files via upload
Statistical forecasting is the art and science of forecasting from data, with or without knowing in advance what equation you should use. The idea is simple: look for statistical patterns in currently available data that you believe will continue into the future. In other words, figure out the way in which the future will look very much like the present, only longer.

Time Series forecasting techniques are very important for a data scientist to master as time-series data occur in every domain from the medical, stock market, climate change prediction, and so on.